Split major functionalities, add basic visual design and connected year-picker which updates the data.
THe data now reacts (and displays on the map) when data points or viewport change. This allows rendering on demand the low-size data (e.g. 0.5, 1 or 3 degrees) and letting the user see what changes.

This commit also adds stub UI for model selection and an idea of how the selection of scenarios might use.

The theme is basic and white to keep it simple, we also use white for buttons because the map itself is colorful and using a button theme coudl be a little distracting (however, as we add more UI elements, it may become necessary)