Tidy up mobile design to be attractive. Bottom legend still needs height. I'm liking this gradient-white interactive design where the map is fully central! Still need to figure out the header and model selector.
I am following the rules that:

Anything bordering directly on the screen can be sharp-edged (e.g. the bottom gradient) and should ideally use color BG matching the map with white/opacity overlay.

Anything interactive shoudl "live" above the map (in the future with gradient box-shadow). This is how the legend at the bottom and the header (static, unclickable) will be differntiated from the zoom, go-to-location, year selector and model selector items.