SoC refactor for the too-bloated EnhancedClimateMap. Bring the data loading/parsing functions into their own files.
Move TimelineSelector to where it belongs, as part of the InterfaceInputs.

Seperate the header.

As a result, the climate map nolonger concerns or needs to know about the viewing mode. It's easier to edit e.g. the Header content and create different versions for mobile. Still keeping all main data in the main map component for now.

The whole map component is essentailly re-useable in that we can adjust the props for the data input it should load in the future. However, the full page view(and subseuqent use of the whole screen, and z-index levels) means (as in line with the requirements which we asked), we get better design at the cost that only one map can easily be rendered at one point (since otherwise e.g. a second maps Zoom buttons would overlap with the first map)

Change screenshot image to Camera.